Working algorithm for reaching Flag and going back to Base

+++ src/main/java/com/domain/Player.java
@@ -1,5 +1,6 @@
 package com.domain;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
@@ -8,7 +9,7 @@ public class Player {
     private Integer id;
     private String name;
     private Integer maxMovesPerRound;
-    private Integer movesLeft;
+    private Double movesLeft;
     private Integer viewRange;
     private Integer x;
     private Integer y;
